refactor: cria pasta para armazenar reducer de ciclos
